refactor(lib): group logger/utilities/lectures + move blocks
Eliminate all loose files at the src/lib/ root by grouping them into
named folders:

- logger/        clientLogger.ts, workerSafeLogger.ts
- utilities/     weightedSample.ts, subtitles.ts (+ gender/serverUrl/previewSecret from #444 §4)
- lectures/      lectureShape.ts, nirmalaVidya.ts, nirmalaVidyaApi.ts
- locales/       locales.ts -> index.ts (preserves @/lib/locales)
- richEditor/    richEditor.ts -> index.ts; src/blocks/ -> richEditor/blocks/

@/lib/locales and @/lib/richEditor keep resolving (file -> folder index.ts),
so those import sites are untouched. Rewire the rest, including the 11
vi.mock('@/lib/lectures/nirmalaVidyaApi') calls and the @/blocks consumers.
Also corrects three unit tests that imported relocated plugins via relative
paths. Purely structural; no behavior change.

Refs #444
