+++ src/components/admin/ProjectSelector/ProjectSelector.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation'
 import { useMemo, useState } from 'react'
 
 import { useProject } from '@/contexts/ProjectContext'
-import { clientLogger } from '@/lib/clientLogger'
+import { clientLogger } from '@/lib/logger/clientLogger'
 import type { ProjectSlug } from '@/payload-types'
 import { getProjectOptions, getProjectsFromRoles } from '@/plugins/access'
