+++ src/lib/locales/index.ts
