+++ src/lib/richEditor/blocks/ContentIndexBlock.ts
