+++ src/lib/richEditor/blocks/SplashBlock.ts
