+++ src/payload.config.ts
@@ -14,8 +14,8 @@ import { GetPlatformProxyOptions } from 'wrangler'
 
 import { serverEnv } from '@/lib/env'
 import { buildPayloadLocales, DEFAULT_LOCALE } from '@/lib/locales'
+import { createWorkerSafeLogger } from '@/lib/logger/workerSafeLogger'
 import { getServerUrl } from '@/lib/utilities/serverUrl'
-import { createWorkerSafeLogger } from '@/lib/workerSafeLogger'
 import { accessPlugin, bypassPermissions, filterAvailableLocales } from '@/plugins/access'
 import { resendAdapter } from '@/plugins/email/resendAdapter'
 import { openapiEndpointAuth, scalarPlugin } from '@/plugins/openapi'
