+++ tests/int/cleanup-orphaned-media.int.spec.ts
@@ -13,7 +13,7 @@ import { createTestEnvironment } from '../utils/testHelpers'
 
 // Mock the Nirmala Vidya API client — testData.createLecture() triggers the
 // beforeChange hook which calls fetchNirmalaVidyaVideo
-vi.mock('@/lib/nirmalaVidyaApi', async () => {
+vi.mock('@/lib/lectures/nirmalaVidyaApi', async () => {
   const { readFileSync } = await import('fs')
   const { dirname, join } = await import('path')
   const { fileURLToPath } = await import('url')
