+++ tests/int/collections-smoke.int.spec.ts
@@ -19,14 +19,14 @@ import { createTestEnvironment } from '../utils/testHelpers'
 // Lectures collection has a beforeChange hook (populateFromNirmalaVidya) that
 // fetches metadata from the Nirmala Vidya API. Stub it so smoke tests never
 // hit the network.
-vi.mock('@/lib/nirmalaVidyaApi', async (importOriginal) => {
+vi.mock('@/lib/lectures/nirmalaVidyaApi', async (importOriginal) => {
   const { readFileSync } = await import('fs')
   const { dirname, join } = await import('path')
   const { fileURLToPath: toPath } = await import('url')
   const imgBuffer = readFileSync(
     join(dirname(toPath(import.meta.url)), '../files/image-1050x700.jpg'),
   )
-  const original = await importOriginal<typeof import('@/lib/nirmalaVidyaApi')>()
+  const original = await importOriginal<typeof import('@/lib/lectures/nirmalaVidyaApi')>()
   return {
     extractVimeoId: vi.fn(original.extractVimeoId),
     fetchNirmalaVidyaVideo: vi.fn().mockResolvedValue({
