+++ tests/int/wemeditateAppStatus.int.spec.ts
@@ -20,14 +20,14 @@ import { testData } from '../utils/testData'
 import { createTestEnvironment } from '../utils/testHelpers'
 
 // Mock the Nirmala Vidya API client to prevent real network calls when creating lectures.
-vi.mock('@/lib/nirmalaVidyaApi', async (importOriginal) => {
+vi.mock('@/lib/lectures/nirmalaVidyaApi', async (importOriginal) => {
   const { readFileSync } = await import('fs')
   const { dirname, join } = await import('path')
   const { fileURLToPath: toPath } = await import('url')
   const imgBuffer = readFileSync(
     join(dirname(toPath(import.meta.url)), '../files/image-1050x700.jpg'),
   )
-  const original = await importOriginal<typeof import('@/lib/nirmalaVidyaApi')>()
+  const original = await importOriginal<typeof import('@/lib/lectures/nirmalaVidyaApi')>()
   return {
     extractVimeoId: vi.fn(original.extractVimeoId),
     fetchNirmalaVidyaVideo: vi.fn().mockResolvedValue({
